+++ gulp/src/sass/analytics.scss
@@ -515,22 +515,21 @@ path:hover {
 
 /*CSS FOR THE BREADCRUMBS*/
 .breadcrumbs-container {
-  display: inline-block;
   list-style: outside none none;
   padding: 0;
-  margin: 0;
+  margin: 0 0 20px 0;
 }
 .analytics-breadcrumbs {
   display: inline;
   margin-right: 10px;
 }
 .breadcrumb-active, .breadcrumb-active:hover, .breadcrumb-active:focus, .breadcrumb-active:active {
-  background: $light-gray;
+  background: $primary-blue;
   display: inline-block;
   outline: medium none;
-  padding: 10px 10px 10px 30px;
+  padding: 3px 3px 3px 20px;
   text-decoration: none;
-  color: $primary-blue;
+  color: $white;
   position: relative;
 }
 .breadcrumb-active:hover {
@@ -538,13 +537,13 @@ path:hover {
   color: $white;
 }
 .breadcrumb-active:focus {
-  background: $primary-blue;
+  background: $primary-red;
   color: $white;
 }
 .breadcrumb-active::before {
-  border-color: transparent transparent transparent $white;
+  border-color: transparent transparent transparent $light-gray;
   border-style: solid;
-  border-width: 20px 0 20px 20px;
+  border-width: 13px 0 13px 13px;
   box-sizing: border-box;
   content: "";
   height: 0;
@@ -554,9 +553,9 @@ path:hover {
   width: 0;
 }
 .breadcrumb-active::after {
-  border-color: transparent transparent transparent $light-gray;
+  border-color: transparent transparent transparent $primary-blue;
   border-style: solid;
-  border-width: 20px 0 20px 20px;
+  border-width: 13px 0 13px 13px;
   box-sizing: border-box;
   content: "";
   height: 0;
@@ -570,7 +569,7 @@ path:hover {
 	border-left-color: $darkest-gray;
 }
 .breadcrumb-active:focus::after {
-	border-left-color: $primary-blue;
+	border-left-color: $primary-red;
 }
 
 /*CSS FOR THE NEW DIMENSION TABLE*/
@@ -1051,6 +1050,10 @@ color: $primary-blue;
        .filter-switch {
          margin-left: 0;
        }
+
+       .filter-label {
+         font-weight: bold;
+       }
      }
      .tab a {
        color: $white;
@@ -1302,7 +1305,7 @@ color: $primary-blue;
          }
        }
        .tab-content-container {
-         margin: 0 0 50px 0;
+         margin: -10px 0 50px 0;
        }
      }
      #mobile_menu {
